backend/requirements.txt
flask
flask-cors
opencv-python
numpy
Pillow
pyttsx3
ultralytics
torch
